+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   - Common use case - trimming out values from Frigate blueprint like `attachment` or `video` that can result in broken thumbnails for Apple notifications
 ## Generic
 - The nested support for `data` filtering introduced for Mobile Push introduced also for Generic transport
+## Other
+- Version of Supernotify added to archived notification
 ## 1.15.0
 ### New transports
 - Pushover, Telegram and LaMetric, contributed by [@lollox80](https://github.com/lollox80)